% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{iterate<-}
\alias{iterate<-}
\alias{iterate<-,spop,data.frame-method}
\alias{iterate<-,spop-method}
\title{Iterate population}
\usage{
iterate(x) <- value

\S4method{iterate}{spop,data.frame}(x) <- value
}
\arguments{
\item{x}{spop class instant}

\item{value}{\code{data.frame} with the following fields
\itemize{
\item \code{death}: age-independent daily probability of death
\item \code{death_mean} and \code{death_sd}: age-dependent daily probability of death (\code{death_sd=0} indicates fixed life time (defined by \code{death_mean}))
\item \code{dev}: age-independent daily probability of development
\item \code{dev_mean} and \code{dev_sd}: age-dependent daily probability of development (\code{dev_sd=0} indicates fixed development time (defined by \code{dev_mean}))
}}
}
\description{
Iterate the population for one day
}
